Added push constants for model, view, and proj matrices. Updated math to use floats instead of doubles :(

+++ src/rtEMath/rtEMath.c
@@ -2,12 +2,12 @@
 #include <math.h>
 #include <string.h>
 
-void rtEMath_vec3MultScalar(vec3 a, double factor, vec3 dest) {
+void rtEMath_vec3MultScalar(vec3 a, float factor, vec3 dest) {
         dest[0] = a[0]*factor;
         dest[1] = a[1]*factor;
         dest[2] = a[2]*factor;
 }
-void rtEMath_vec3DivScalar(vec3 a, double divisor, vec3 dest) {
+void rtEMath_vec3DivScalar(vec3 a, float divisor, vec3 dest) {
         dest[0] = a[0]/divisor;
         dest[1] = a[1]/divisor;
         dest[2] = a[2]/divisor;
@@ -27,14 +27,14 @@ void rtEMath_vec3Cross(vec3 a, vec3 b, vec3 dest) {
         dest[1] = a[2]*b[0] - a[0]*b[2];
         dest[2] = a[0]*b[1] - a[1]*b[0];
 }
-double rtEMath_vec3Dot(vec3 a, vec3 b) {
+float rtEMath_vec3Dot(vec3 a, vec3 b) {
         return a[0]*b[0] + a[1]*b[1] + a[2]*b[2];
 }
 void rtEMath_vec3Normalize(vec3 a, vec3 dest) {
         rtEMath_vec3DivScalar(a, rtEMath_vec3GetMagnitude(a), dest);
 }
 
-double rtEMath_vec3GetMagnitude(vec3 a) {
+float rtEMath_vec3GetMagnitude(vec3 a) {
         return sqrt(
                         a[0]*a[0] +
                         a[1]*a[1] +
@@ -50,7 +50,7 @@ void rtEMath_mat4CreateModel(vec3 a, mat4 dest) {
                 a[0],a[1],a[2], 1.0
         };
         
-        memcpy(dest, tmpModel, sizeof(double) * 16);
+        memcpy(dest, tmpModel, sizeof(float) * 16);
 }
 void rtEMath_mat4CreateLookAt(vec3 position, vec3 target, vec3 up, mat4 dest) {
         vec3 invDirection = RTEMATH_VEC3_ZERO; 
@@ -76,25 +76,25 @@ void rtEMath_mat4CreateLookAt(vec3 position, vec3 target, vec3 up, mat4 dest) {
                 0.0, 0.0, 0.0, 1.0
         };
 
-        memcpy(dest, tmpMat, sizeof(double) * 16);
+        memcpy(dest, tmpMat, sizeof(float) * 16);
 }
 
-void rtEMath_mat4CreatePerspectiveProjection(double fov, double near, double far, double aspect, mat4 dest) {
+void rtEMath_mat4CreatePerspectiveProjection(float fov, float near, float far, float aspect, mat4 dest) {
         // big thanks to https://learnwebgl.brown37.net/08_projections/projections_perspective.html
-        double top = near * tan(fov/2);
-        double bottom = -top;
-        double right = top*aspect;
-        double left = -top;
+        float top = near * tan(fov/2);
+        float bottom = -top;
+        float right = top*aspect;
+        float left = -top;
 
-        double midX = (left + right)/2.0;
-        double midY = (bottom + top)/2.0;
+        float midX = (left + right)/2.0;
+        float midY = (bottom + top)/2.0;
 
-        double scaleX = 2.0 / (right - left);
-        double scaleY = 2.0 / (top - bottom);
+        float scaleX = 2.0 / (right - left);
+        float scaleY = 2.0 / (top - bottom);
 
         
-        double c1 = (2*far*near) / (near - far);
-        double c2 = (far + near) / (far - near);
+        float c1 = (2*far*near) / (near - far);
+        float c2 = (far + near) / (far - near);
 
         mat4 tmpMat = {
                 near*scaleX, 0          , 0 ,  0,
@@ -103,7 +103,7 @@ void rtEMath_mat4CreatePerspectiveProjection(double fov, double near, double far
                 midX       , midY       , c2,  0
         };
 
-        memcpy(dest, tmpMat, sizeof(double) * 16);
+        memcpy(dest, tmpMat, sizeof(float) * 16);
 }
